+++ tests/dummy/app/items/item/edit/route.js
@@ -5,9 +5,19 @@ export default Ember.Route.extend({
   model (params) {
     let item = this.modelFor('items.item');
     Ember.debug('Items.item.edit id: ' + params.id + ' item.id ' + item.id + JSON.stringify(params));
-    return Ember.RSVP.hash({
-      item: item,
-      data: this.get('itemsService').getDataById(item.id)
-    });
+
+    // only get the data if this is a type with data!
+    let validTypes = ['Web Mapping Application', 'Web Map'];
+    if (validTypes.includes(item.type)) {
+      return Ember.RSVP.hash({
+        item: item,
+        data: this.get('itemsService').getDataById(item.id)
+      });
+    } else {
+      return Ember.RSVP.resolve({
+        item: item,
+        data: {}
+      });
+    }
   }
 });